\name{convert_na}
\alias{convert_na}
\title{Convert values with specific comment or color to NAs}
\usage{
convert_na(this.dir, sheet.name, chose.type, chose.comment, chose.color)
}
\description{
The function is used to convert values with specific comments or colors to NAs in excel files which is produced by MetIDQ software. MetIDQ is used to identify and quantify metabolites for LC-MS or FIA-MS.
}
\examples{
}
